[Fix] DailyFragment 이미지 오류 수정

+++ app/src/main/java/com/boostcampai/foodlog/Utils.kt
@@ -14,8 +14,8 @@ fun convertBitmapToBase64(bitmap: Bitmap): String {
     val height = bitmap.height
     val resized = Bitmap.createScaledBitmap(
         bitmap,
-        width.toInt(),
-        height.toInt(),
+        width,
+        height,
         false
     )
     resized.compress(Bitmap.CompressFormat.JPEG, 100, outputStream)
